+++ src/components/systemConfig/kernel/Kernel.d.ts
@@ -1,0 +1,5 @@
+export interface Kernel {
+  id: string;
+  name: string;
+  isInstalled:boolean;
+}
