+++ src/globals.d.ts
@@ -1,1 +1,2 @@
 declare module '*.md';
+declare module '*.d.ts';
