--- src/stores/FastestMirrorStore.ts
@@ -1,12 +1,0 @@
-import {
-  atom,
-} from 'recoil';
-
-export interface MirrorProp {
-  isProcessing: boolean,
-}
-
-export const mirrorState = atom({
-  key: 'mirrorState',
-  default: { isProcessing: false } as MirrorProp,
-});
